[refactor] App: move interfaces to dedicated package
