+++ composeApp/src/commonMain/kotlin/uk/ryanwong/gmap2ics/data/datasources/local/interfaces/LocalDataSource.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2022-2024. Ryan Wong (hello@ryanwebmail.com)
  */
 
-package uk.ryanwong.gmap2ics.data.datasources.local
+package uk.ryanwong.gmap2ics.data.datasources.local.interfaces
 
 interface LocalDataSource {
     suspend fun getFileList(absolutePath: String, extension: String): Result<List<String>>
