+++ composeApp/src/commonMain/kotlin/uk/ryanwong/gmap2ics/data/repositories/LocalFileRepositoryImpl.kt
@@ -4,8 +4,8 @@
 
 package uk.ryanwong.gmap2ics.data.repositories
 
-import uk.ryanwong.gmap2ics.data.datasources.local.LocalDataSource
-import uk.ryanwong.gmap2ics.data.datasources.local.impl.LocalDataSourceImpl
+import uk.ryanwong.gmap2ics.data.datasources.local.interfaces.LocalDataSource
+import uk.ryanwong.gmap2ics.data.datasources.local.LocalDataSourceImpl
 import uk.ryanwong.gmap2ics.domain.models.VEvent
 import uk.ryanwong.gmap2ics.domain.repositories.LocalFileRepository
 import java.nio.file.Paths
